cleanup: move fetch and url helpers to utils, improve some typing
